+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers.NUnit.Integration/ArtifactHelpers/DocumentHelperIntegrationTests.cs
@@ -14,9 +14,6 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 		private string _workspaceName;
 		private IServicesMgr _servicesManager;
 		private int _workspaceId;
-		private IDBContext _dbContext;
-		private KeplerHelper _keplerHelper;
-		private bool useDbContext;
 
 		[SetUp]
 		public void SetUp()
@@ -30,14 +27,6 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 			_workspaceName = $"IntTest_{Guid.NewGuid()}";
 			_servicesManager = testHelper.GetServicesManager();
 			_workspaceId = Helpers.WorkspaceHelpers.WorkspaceHelpers.CreateAsync(_servicesManager, _workspaceName, SharedTestHelpers.ConfigurationHelper.TEST_WORKSPACE_TEMPLATE_NAME).ConfigureAwait(false).GetAwaiter().GetResult();
-			_keplerHelper = new KeplerHelper();
-
-			bool isKeplerCompatible = _keplerHelper.IsVersionKeplerCompatibleAsync().ConfigureAwait(false).GetAwaiter().GetResult();
-			useDbContext = !isKeplerCompatible || ConfigurationHelper.FORCE_DBCONTEXT.Trim().ToLower().Equals("true");
-			if (useDbContext)
-			{
-				_dbContext = testHelper.GetDBContext(_workspaceId);
-			}
 		}
 
 		[TearDown]
@@ -47,74 +36,23 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 			Helpers.WorkspaceHelpers.WorkspaceHelpers.Delete(_servicesManager, _workspaceId);
 
 			testHelper = null;
-			_dbContext = null;
 			_servicesManager = null;
-			_dbContext = null;
 		}
 
-		[Test]
-		public void GetDocumentIdentifierFieldColumnName()
-		{
-			// Arrange
-			const int fieldArtifactTypeId = 10;
-			const string controlNumber = "ControlNumber";
-
-			// Act
-			string columnName = "";
-			if (useDbContext)
-			{
-				columnName = Document.GetDocumentIdentifierFieldColumnName(_dbContext, fieldArtifactTypeId);
-			}
-			else
-			{
-				columnName = Document.GetDocumentIdentifierFieldColumnName(fieldArtifactTypeId, _workspaceId, _keplerHelper);
-			}
-
-			// Assert
-			Assert.AreEqual(controlNumber, columnName);
-		}
 
 		[Test]
 		public void GetDocumentIdentifierFieldName()
 		{
 			// Arrange
-			const int fieldArtifactTypeId = 10;
+			const int fieldArtifactTypeId = 14;
 			const string controlNumber = "Control Number";
 
 			// Act
 			string fieldName = "";
-			if (useDbContext)
-			{
-				fieldName = Document.GetDocumentIdentifierFieldName(_dbContext, fieldArtifactTypeId);
-			}
-			else
-			{
-				fieldName = Document.GetDocumentIdentifierFieldName(fieldArtifactTypeId, _workspaceId, _keplerHelper);
-			}
+			fieldName = Document.GetDocumentIdentifierFieldName(_servicesManager, _workspaceId, fieldArtifactTypeId);
 
 			// Assert
 			Assert.AreEqual(controlNumber, fieldName);
-		}
-
-		[Test]
-		public void GetDocumentIdentifierFieldColumnName_InvalidFieldType()
-		{
-			// Arrange
-			const int fieldArtifactTypeId = 0;
-
-			// Act
-			string columnName = "";
-			if (useDbContext)
-			{
-				columnName = Document.GetDocumentIdentifierFieldColumnName(_dbContext, fieldArtifactTypeId);
-			}
-			else
-			{
-				columnName = Document.GetDocumentIdentifierFieldColumnName(fieldArtifactTypeId, _workspaceId, _keplerHelper);
-			}
-
-			// Assert
-			Assert.AreEqual(null, columnName);
 		}
 
 		[Test]
@@ -123,19 +61,10 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 			// Arrange
 			const int fieldArtifactTypeId = 0;
 
-			// Act
+			// Act / Assert
 			string fieldName = "";
-			if (useDbContext)
-			{
-				fieldName = Document.GetDocumentIdentifierFieldName(_dbContext, fieldArtifactTypeId);
-			}
-			else
-			{
-				fieldName = Document.GetDocumentIdentifierFieldName(fieldArtifactTypeId, _workspaceId, _keplerHelper);
-			}
-
-			// Assert
-			Assert.AreEqual(null, fieldName);
+			Assert.Throws<Exception>(() =>
+				Document.GetDocumentIdentifierFieldName(_servicesManager, _workspaceId, fieldArtifactTypeId));
 		}
 	}
 }